Use Rust slices
Instead of passing around the reference to the whole vector of digits
and the start index separately, make use of a Rust idiom: slices.

We can pass around a &[u8], which is a slice of the original vector.
We can slice it further, such as passing &digits[i+1..] to a recursive
invocation.