Write directly to the output stream
Using a formatter to print the output is much slower than writing
directly to the output stream.

We only need to write things that have a fixed format, and therefore
writing directly to the output stream speeds things up considerably -
on my machine I get a 3-4x speed up in the benchmarks.